Move activity list to separate component enclosed in Suspense
